Crate oauth2_types

OAuth 2.0 and OpenID Connect types.

This is part of the Matrix Authentication Service project.

Modules

errors

Error types returned by an authorization server.

oidc

Types to interact with the OpenID Connect specification.

pkce

Types for the Proof Key for Code Exchange.

prelude

Traits intended for blanket imports.

registration

Types for Dynamic Client Registration.

requests

Requests and response types to interact with the OAuth 2.0 specification.

response_type

Response types in the OpenID Connect specification.

scope

Types to define an access token’s scope.

webfinger

Types for provider discovery using Webfinger.