Struct oauth2_types::requests::AuthorizationRequest

pub struct AuthorizationRequest {

    pub response_type: ResponseType,
    pub client_id: String,
    pub redirect_uri: Option<Url>,
    pub scope: Scope,
    pub state: Option<String>,
    pub response_mode: Option<ResponseMode>,
    pub nonce: Option<String>,
    pub display: Option<Display>,
    pub prompt: Option<Vec<Prompt>>,
    pub max_age: Option<NonZeroU32>,
    pub ui_locales: Option<Vec<LanguageTag>>,
    pub id_token_hint: Option<String>,
    pub login_hint: Option<String>,
    pub acr_values: Option<HashSet<String>>,
    pub request: Option<String>,
    pub request_uri: Option<Url>,
    pub registration: Option<String>,

}

The body of a request to the Authorization Endpoint.

Fields
response_type: ResponseType

OAuth 2.0 Response Type value that determines the authorization processing flow to be used.

client_id: String

OAuth 2.0 Client Identifier valid at the Authorization Server.

redirect_uri: Option<Url>

Redirection URI to which the response will be sent.

This field is required when using a response type returning an authorization code.

This URI must have been pre-registered with the OpenID Provider.

scope: Scope

The scope of the access request.

OpenID Connect requests must contain the openid scope value.

state: Option<String>

Opaque value used to maintain state between the request and the callback.

response_mode: Option<ResponseMode>

The mechanism to be used for returning parameters from the Authorization Endpoint.

This use of this parameter is not recommended when the Response Mode that would be requested is the default mode specified for the Response Type.

nonce: Option<String>

String value used to associate a Client session with an ID Token, and to mitigate replay attacks.

display: Option<Display>

How the Authorization Server should display the authentication and consent user interface pages to the End-User.

prompt: Option<Vec<Prompt>>

Whether the Authorization Server should prompt the End-User for reauthentication and consent.

If Prompt::None is used, it must be the only value.

max_age: Option<NonZeroU32>

The allowable elapsed time in seconds since the last time the End-User was actively authenticated by the OpenID Provider.

ui_locales: Option<Vec<LanguageTag>>

End-User’s preferred languages and scripts for the user interface.

id_token_hint: Option<String>

ID Token previously issued by the Authorization Server being passed as a hint about the End-User’s current or past authenticated session with the Client.

login_hint: Option<String>

Hint to the Authorization Server about the login identifier the End-User might use to log in.

acr_values: Option<HashSet<String>>

Requested Authentication Context Class Reference values.

request: Option<String>

A JWT that contains the request’s parameter values, called a Request Object.

request_uri: Option<Url>

A URI referencing a Request Object or a Pushed Authorization Request.

registration: Option<String>

A JSON object containing the Client Metadata when interacting with a Self-Issued OpenID Provider.

Implementations

impl AuthorizationRequest

pub fn new(response_type: ResponseType, client_id: String, scope: Scope) -> Self

Creates a basic AuthorizationRequest.